DB-9: product folder structure and props
