+++ src/data/Pdata.js
@@ -9,12 +9,15 @@ const Pdata = {
       brand: 'Nike',
       productRating: 3.5,
       numReviewsThreshold: 10,
-      numOfReviews: 50,
+      numberOfSold: 65,
       discount: 'ETB7,632.84',
       extraDiscount: 2,
       discountPercent: 40,
       dealType: 'Wellcome deal',
-      discountType: 'Choice Day',
+      topSoldItem: 'Top seling',
+      date: 5,
+      choice: 'Choice',
+      freeShip: 2000,
       description: 'A popular shirt for new comers',
     },
     {
@@ -26,9 +29,10 @@ const Pdata = {
       brand: 'Nike',
       productRating: 4.5,
       numReviewsThreshold: 10,
-      numberOfSold: 65,
+      numberOfSold: 3,
       discount: 'ETB8,271.34',
-      discountType: 'Choice Day',
+      choice: 'Choice',
+      freeShip: 5000,
       description: 'nice brand popular shirt for everyone ',
       alt: 'A popular shirt',
     },
@@ -47,7 +51,8 @@ const Pdata = {
       discountPercent: 25,
       dealType: 'Wellcome deal',
       discount: 'ETB5,453.34',
-
+      topSoldItem: 'Top seling',
+      date: 7,
       description: 'brand new comming t shirt for us',
       alt: 'A popular shirt',
     },
@@ -63,6 +68,8 @@ const Pdata = {
       numOfReviews: 3,
       numOfSold: 200,
       discount: 'ETB3,759.28',
+      choice: 'Choice',
+      freeShip: 5000,
       description: 'pants for boys quality added',
       alt: 'A popular fatty shirt',
     },
@@ -81,6 +88,8 @@ const Pdata = {
       extraDiscount: 2,
       discountPercent: 50,
       dealType: 'Wellcome deal',
+      topSoldItem: 'Top seling',
+      date: 3,
       description: 'likely and lovely pant form differnet purpose',
       alt: 'A popular pant',
     },
@@ -96,6 +105,8 @@ const Pdata = {
       numOfReviews: 50,
       numOfSold: 150,
       discount: 'ETB9,753.83',
+      choice: 'Choice',
+      freeShip: 5000,
       description: 'Product Specification And Titles For The User',
       alt: 'A popular pant',
     },
@@ -128,6 +139,10 @@ const Pdata = {
       numOfReviews: 50,
       numOfSold: 150,
       discount: 'ETB9,753.83',
+      topSoldItem: 'Top seling',
+      choice: 'Choice',
+      freeShip: 5000,
+      date: 8,
       description: 'pants for long living',
       alt: 'A popular pant',
     },
