+++ src/data/product.tsx
@@ -1,0 +1,23 @@
+const product = {
+  name: 'Free Shirt',
+  catagory: 'Shirt',
+  image: '/assets/images/product.webp',
+  alt: 'product image',
+  price: 7392.45,
+  oldPrice: 5643.43,
+  brand: 'Nike',
+  productRating: 3.5,
+  numReviewsThreshold: 10,
+  numberOfSold: 65,
+  discount: 'ETB7,632.84',
+  extraDiscount: 2,
+  discountPercent: 40,
+  dealType: 'Wellcome deal',
+  topSoldItem: 'Top seling',
+  date: 5,
+  choice: 'Choice',
+  freeShip: 2000,
+  description: 'A popular shirt for new comers',
+};
+
+export default product;
