+++ src/features/ProductList/components/AddToCartButton.tsx
@@ -2,33 +2,40 @@ import { IconButton } from '@mui/material';
 import React from 'react';
 import AddShoppingCartRoundedIcon from '@mui/icons-material/AddShoppingCartRounded';
 
-export default function CartingButton() {
+interface AddToCartButton {
+  id: string;
+}
+
+export default function AddToCartButton({ id }: AddToCartButton) {
   return (
-    <div>
+    <>
       <IconButton
+        id={id}
         size="large"
         aria-label="add to shopping cart"
+        color="primaryButton"
         sx={{
           position: 'absolute',
-          color: 'black',
+          // color: 'black',
           bottom: '10%',
           right: '10%',
           backgroundColor: 'background.paper',
           '&:hover': {
-            background: 'black',
+            // background: 'black',
             '.cart-icon': { color: 'white' },
           },
         }}
       >
         <AddShoppingCartRoundedIcon
           className="cart-icon"
+          // color=""
           sx={{
             '&:hover': {
-              color: (theme) => theme.palette.brighten.main,
+              // color: (theme) => theme.palette.brighten.main,
             },
           }}
         />
       </IconButton>
-    </div>
+    </>
   );
 }
