+++ src/features/ProductList/components/SellingPrice.tsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import { Box, Stack, Typography } from '@mui/material';
+import { Typography } from '@mui/material';
 import getPrice from '@/utils/getPrice';
 
 interface priceProps {
@@ -7,7 +7,7 @@ interface priceProps {
   oldPrice: number;
 }
 
-export default function Pricing({ price, oldPrice }: priceProps) {
+export default function SellingPrice({ price, oldPrice }: priceProps) {
   const { int, dec } = getPrice(price);
 
   return (
@@ -20,8 +20,8 @@ export default function Pricing({ price, oldPrice }: priceProps) {
             fontSize: '.75rem',
             fontWeight: 'bold',
           },
-          '.price': {
-            fontSize: '1.75rem',
+          '.price-int': {
+            fontSize: '1.5rem',
             fontWeight: 'bold',
           },
           '.price-dec': {
@@ -31,7 +31,7 @@ export default function Pricing({ price, oldPrice }: priceProps) {
         }}
       >
         <span className="currency">ETB</span>
-        <span className="price">{int}</span>
+        <span className="price-int">{int}</span>
         <span className="price-dec">.{dec}</span>
       </Typography>
